ML
Did a lot of changes into the ML part. Is almost 9 and im not sure when im gonna pass out so im sending this in case i dont have enough time for the last model to finish train. After that i plan to use a soft voting classifier to improve the performance. Easy to implement if find problems with it let me know